Refactor and improve Python server
This commit introduces a wide range of corrections, improvements, and modernizations to the Python gesture server.

Corrections:
- Renamed the directory `Server/Pyhhon` to `Server/Python`.
- Renamed `gesture_serveur.py` to `gesture_server.py`.
- Fixed a buggy gesture prediction algorithm.
- Fixed a bug where two performance monitors were used, leading to incorrect reporting.

Improvements:
- The server is now fully non-blocking. All blocking `pyautogui` calls have been moved to a `ThreadPoolExecutor` to prevent stalling the main `asyncio` event loop.
- Added an `asyncio.Lock` to protect shared state in the `GestureExecutor`, making it thread-safe.
- Implemented robust error handling in TCP and WebSocket handlers.
- Simplified and corrected the logging configuration.
- Removed version pins from `requirements.txt` to allow for more flexible dependency resolution, and removed the unused and unavailable `tensorflow-lite` package.

New Features:
- Added a `config.yaml` file and implemented a resilient configuration loader that falls back to default values.
- Added a `test_client.py` script to facilitate testing of the WebSocket server.
